refactor: pull out nav and entity browser
